feat(@angular-devkit/core): invalid smart default sources are now returning undefined
This is a break from the previous API, but it wasnt properly defined. Smart defaults are
never guaranteed to return a value anyway.

The next step in cleaning up smart defaults would be to have them run _before_ validation
so that smart defaults should be validated for type and optional/require.